chore: refactor relative time function
